Day 2!
All the opcoding in Python for AoC Day 2. Today I learned way too much and I barely was able to put this together lol. Opcodes were easy to understand, however, the way to read and write to a list was the hardest part to get to. Hoping one day I'll get the understanding in order to do this without as much assistance (shout out to my friend)!